pyproject.toml
[project]
name = "llm-engineering-practice"
version = "0.1.0"
description = "Hands-on practice with LLM engineering concepts — prompting, tokenization, memory, agents, and more."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "google-genai>=1.0.0",
    "gradio>=4.0.0",
    "huggingface-hub>=0.19.3",
    "ipython>=8.18.1",
    "litellm>=1.0.0",
    "openai>=1.16.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tiktoken>=0.7.0",
]

[project.scripts]
start = "main:main"
